#443: Moved the PassiveEquipmentCategoryUnlockParser to the model folder
 -
