+++ src/app/ffbe/mappers/effects/passive-effect-parser.factory.ts
@@ -1,6 +1,6 @@
 import {EffectParser} from './effect-parser';
 import {UnknownEffectParser} from './unknown-effect-parser';
-import {PassiveEquipmentCategoryUnlockParser} from './passives/passive-equipment-category-unlock.parser';
+import {PassiveEquipmentCategoryUnlockParser} from '../../model/effects/passives/passive-equipment-category-unlock.parser';
 import {PassiveStatsIncreaseHpThresholdParser} from './passives/passive-stats-increase-hp-threshold.parser';
 import {PassiveEquipmentCategoryStatsIncreaseParser} from './passives/passive-equipment-category-stats-increase.parser';
 import {PassiveCoverParser} from './passives/passive-cover.parser';
